Bot now ignores pinned messages

+++ bot.py
@@ -44,7 +44,7 @@ async def chomp(interaction: discord.Interaction, rate: int = 1):
             CHOMPING=True
             async for message in channel.history(limit=None, oldest_first=False):
                 if (not CHOMPING): break
-                if ((datetime.utcnow().replace(tzinfo=None) - message.created_at.replace(tzinfo=None) > timedelta(minutes=1)) or (message.author != BOT.user)):
+                if (((datetime.utcnow().replace(tzinfo=None) - message.created_at.replace(tzinfo=None) > timedelta(minutes=1)) or (message.author != BOT.user)) and (not message.pinned)):
                     print("Chomped! -> " + message.content)
                     await message.delete()
                     await asyncio.sleep(1.0/rate)  # Avoid rate limits
